+++ Dev/Editor/EffekseerCore/Data/Value/FCurveColor.cs
@@ -8,18 +8,18 @@ namespace Effekseer.Data.Value
 	public class FCurveColorRGBA
 	{
 		public Value.Enum<FCurveTimelineType> Timeline = new Enum<FCurveTimelineType>();
-		public FCurve<byte> R { get; private set; }
-		public FCurve<byte> G { get; private set; }
-		public FCurve<byte> B { get; private set; }
-		public FCurve<byte> A { get; private set; }
+		public FCurve<int> R { get; private set; }
+		public FCurve<int> G { get; private set; }
+		public FCurve<int> B { get; private set; }
+		public FCurve<int> A { get; private set; }
 
 		public FCurveColorRGBA()
 		{
 			Timeline = new Enum<FCurveTimelineType>();
-			R = new FCurve<byte>(255);
-			G = new FCurve<byte>(255);
-			B = new FCurve<byte>(255);
-			A = new FCurve<byte>(255);
+			R = new FCurve<int>(255);
+			G = new FCurve<int>(255);
+			B = new FCurve<int>(255);
+			A = new FCurve<int>(255);
 		}
 
 		public byte[] GetBytes(float mul = 1.0f)
